refactor(analysis): extract host service logs (hostlog) into its own module (#416)
Sixth slice of the analysis decomposition. The per-host service-log
extractors were already an SPI (List<HostServiceLogExtractor>); promote
that seam into analysis.spi and move the implementations out.

- SPI contract -> com.tracepcap.analysis.spi: HostServiceLogExtractor
  (interface), HostServiceLogResult, HostServiceSuspicion, and a new
  ServiceLogRoles holding the shared dns/web/api role vocabulary
- Implementations + storage -> com.tracepcap.hostlog: DnsQueryLogExtractor,
  WebServerLogExtractor (service); DnsQueryLog/HttpEndpointLog entities
  (entity); their repositories (repository); extractor tests
- Decouple the core->feature leak: AnalysisService used
  DnsQueryLogExtractor.ROLE, and the classifier signals + intelligence
  used the extractors' ROLE/ROLE_WEB/ROLE_API constants. All now use
  ServiceLogRoles from the SPI, so nothing outside hostlog depends on the
  concrete extractors. intelligence still reads the log entities (a
  legitimate feature->feature data dependency).

OpenAPI contract unchanged. Verified with mvn verify (JUnit +
Testcontainers).

Part of #416.
