+++ backend/src/main/java/com/tracepcap/analysis/spi/HostServiceLogResult.java
@@ -1,4 +1,4 @@
-package com.tracepcap.analysis.service.hostlog;
+package com.tracepcap.analysis.spi;
 
 import java.util.List;
 import java.util.Map;
