+++ backend/src/main/java/com/tracepcap/analysis/spi/HostServiceSuspicion.java
@@ -1,4 +1,4 @@
-package com.tracepcap.analysis.service.hostlog;
+package com.tracepcap.analysis.spi;
 
 /**
  * A server host flagged as suspicious by a {@link HostServiceLogExtractor}.
